diabetes_prediction/requirements.txt
# Django and extensions
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
djangorestframework==3.14.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.1
joblib==1.3.2
pandas==2.1.1
numpy==1.26.0

# OCR and PDF processing
pytesseract==0.3.10
opencv-python==4.8.1.78
PyMuPDF==1.23.5
Pillow==10.1.0

# Web development
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2